migrate json rpc config
