+++ src/config/mod.rs
@@ -3,29 +3,36 @@
 //! Reads a TOML config file for the IPC Agent and deserializes it in a type-safe way into a
 //! [`Config`] struct.
 
+mod deserialize;
+
 use std::collections::HashMap;
-use std::fmt::Formatter;
 use std::fs;
-use std::str::FromStr;
+use std::net::SocketAddr;
 
 use anyhow::Result;
 use fvm_shared::address::Address;
 use ipc_sdk::subnet_id::SubnetID;
-use serde::de::{Error, SeqAccess};
-use serde::{Deserialize, Deserializer};
+use serde::Deserialize;
 use url::Url;
+use crate::config::deserialize::{deserialize_subnet_id, deserialize_accounts};
 
 /// The top-level struct representing the config. Calls to [`Config::from_file`] deserialize into
 /// this struct.
 #[derive(Deserialize)]
 pub(crate) struct Config {
+    pub server: Server,
     pub subnets: HashMap<String, Subnet>,
+}
+
+#[derive(Deserialize, Clone)]
+pub struct Server {
+    pub json_rpc_address: SocketAddr,
 }
 
 /// Represents a subnet declaration in the config.
 #[derive(Deserialize)]
 pub struct Subnet {
-    #[serde(deserialize_with = "deserialize_path")]
+    #[serde(deserialize_with = "deserialize_subnet_id")]
     id: SubnetID,
     jsonrpc_api_http: Url,
     jsonrpc_api_ws: Option<Url>,
@@ -49,65 +56,14 @@ impl Config {
     }
 }
 
-/// A serde deserialization method to deserialize a subnet path string into a [`SubnetID`].
-fn deserialize_path<'de, D>(deserializer: D) -> Result<SubnetID, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    struct SubnetIDVisitor;
-    impl<'de> serde::de::Visitor<'de> for SubnetIDVisitor {
-        type Value = SubnetID;
-
-        fn expecting(&self, formatter: &mut Formatter) -> std::fmt::Result {
-            formatter.write_str("a string")
-        }
-
-        fn visit_str<E>(self, v: &str) -> std::result::Result<Self::Value, E>
-        where
-            E: Error,
-        {
-            SubnetID::from_str(v).map_err(E::custom)
-        }
-    }
-    deserializer.deserialize_str(SubnetIDVisitor)
-}
-
-/// A serde deserialization method to deserialize a list of account strings into a vector of
-/// [`Address`].
-fn deserialize_accounts<'de, D>(deserializer: D) -> Result<Vec<Address>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    struct AddressSeqVisitor;
-    impl<'de> serde::de::Visitor<'de> for AddressSeqVisitor {
-        type Value = Vec<Address>;
-
-        fn expecting(&self, formatter: &mut Formatter) -> std::fmt::Result {
-            formatter.write_str("a sequence of strings")
-        }
-
-        fn visit_seq<A>(self, mut seq: A) -> std::result::Result<Self::Value, A::Error>
-        where
-            A: SeqAccess<'de>,
-        {
-            let mut vec: Vec<Address> = Vec::new();
-            while let Some(value) = seq.next_element::<String>()? {
-                let a = Address::from_str(value.as_str()).map_err(Error::custom)?;
-                vec.push(a);
-            }
-            Ok(vec)
-        }
-    }
-    deserializer.deserialize_str(AddressSeqVisitor)
-}
-
 #[cfg(test)]
 mod tests {
+    use std::net::SocketAddr;
     use std::str::FromStr;
 
     use fvm_shared::address::Address;
     use indoc::formatdoc;
-    use ipc_sdk::subnet_id::{SubnetID, ROOTNET_ID};
+    use ipc_sdk::subnet_id::{ROOTNET_ID, SubnetID};
     use url::Url;
 
     use crate::config::Config;
@@ -115,6 +71,8 @@ mod tests {
     #[test]
     fn read_config() {
         // Arguments for the config's fields
+        let server_json_rpc_addr = "127.0.0.1:3030";
+
         let root_id = "/root";
         let child_id = "/root/f0100";
         let root_auth_token = "root_auth_token";
@@ -125,6 +83,9 @@ mod tests {
 
         let config_str = formatdoc!(
             r#"
+            [server]
+            json_rpc_address = "{server_json_rpc_addr}"
+
             [subnets]
 
             [subnets.root]
@@ -143,6 +104,11 @@ mod tests {
 
         println!("{}", config_str);
         let config = Config::from_str(config_str.as_str()).unwrap();
+
+        assert_eq!(
+            config.server.json_rpc_address,
+            SocketAddr::from_str(server_json_rpc_addr).unwrap()
+        );
 
         let root = &config.subnets["root"];
         assert_eq!(root.id, *ROOTNET_ID);
